make licenses screen part of settings, remove LicenseActivity
