+++ app/src/main/AndroidManifest.xml
@@ -250,14 +250,6 @@
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="com.nutomic.syncthingandroid.activities.DeviceActivity" />
         </activity>
-        <activity android:name=".activities.LicenseActivity"
-            android:label="@string/open_source_licenses_title"
-            android:exported="false"
-            android:parentActivityName=".activities.SettingsActivity">
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value=".activities.SettingsActivity" />
-        </activity>
         <property
             android:name="android.content.MEDIA_CAPABILITIES"
             android:resource="@xml/media_capabilities" />
