--- app/src/main/java/com/nutomic/syncthingandroid/activities/LicenseActivity.kt
@@ -1,89 +1,0 @@
-package com.nutomic.syncthingandroid.activities
-
-import android.os.Bundle
-
-import androidx.activity.compose.LocalOnBackPressedDispatcherOwner
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.Icon
-import androidx.compose.material3.LargeTopAppBar
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBarDefaults
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.input.nestedscroll.nestedScroll
-import androidx.compose.ui.platform.LocalResources
-import androidx.compose.ui.res.stringResource
-import androidx.core.view.WindowCompat
-
-import com.mikepenz.aboutlibraries.ui.compose.produceLibraries
-import com.mikepenz.aboutlibraries.ui.compose.m3.LibrariesContainer
-
-import com.nutomic.syncthingandroid.R
-import com.nutomic.syncthingandroid.theme.ApplicationTheme
-
-class LicenseActivity : ThemedAppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        // Opt-in to edge-to-edge
-        WindowCompat.setDecorFitsSystemWindows(window, false)
-
-        setContent {
-            LicenseScreen()
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun LicenseScreen() {
-    val scrollBehavior = TopAppBarDefaults.exitUntilCollapsedScrollBehavior()
-
-    ApplicationTheme {
-        Surface(modifier = Modifier.fillMaxSize()) {
-            val backDispatcher = LocalOnBackPressedDispatcherOwner.current?.onBackPressedDispatcher
-            val resources = LocalResources.current
-
-            // Read the libraries content outside of produceLibraries to avoid the lint warning
-            val librariesContent = remember {
-                resources.openRawResource(R.raw.aboutlibraries).bufferedReader().use { it.readText() }
-            }
-            val libraries by produceLibraries(librariesContent)
-
-            Scaffold(
-                modifier = Modifier.nestedScroll(scrollBehavior.nestedScrollConnection),
-                topBar = {
-                    LargeTopAppBar(
-                        title = { Text(stringResource(id = R.string.open_source_licenses_title)) },
-                        navigationIcon = {
-                            IconButton(onClick = { backDispatcher?.onBackPressed()  }) {
-                                Icon(
-                                    imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                                    contentDescription = stringResource(id = R.string.back)
-                                )
-                            }
-                        },
-                        scrollBehavior = scrollBehavior,
-                    )
-                }
-            ) { paddingValues ->
-                LibrariesContainer(
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .padding(paddingValues),
-                    libraries = libraries
-                )
-            }
-        }
-    }
-}
